Исправление по ревью
